+++ spring-boot-project/spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/neo4j/Neo4jProperties.java
@@ -19,19 +19,9 @@ package org.springframework.boot.autoconfigure.neo4j;
 import java.io.File;
 import java.net.URI;
 import java.time.Duration;
-import java.util.Locale;
-import java.util.concurrent.TimeUnit;
 
-import org.neo4j.driver.AuthToken;
-import org.neo4j.driver.AuthTokens;
-import org.neo4j.driver.Config;
-import org.neo4j.driver.internal.Scheme;
 import org.neo4j.driver.net.ServerAddressResolver;
-
-import org.springframework.beans.BeanUtils;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.source.InvalidConfigurationPropertyValueException;
-import org.springframework.util.StringUtils;
 
 /**
  * Used to configure an instance of the {@link org.neo4j.driver.Driver Neo4j-Java-Driver}.
@@ -40,21 +30,21 @@ import org.springframework.util.StringUtils;
  * @since 2.4.0
  */
 @ConfigurationProperties(prefix = "spring.neo4j")
-public class Neo4jDriverProperties {
+public class Neo4jProperties {
 
 	/**
-	 * The uri this driver should connect to. The driver supports bolt or neo4j as
-	 * schemes. Defaults to bolt://localhost:7687.
+	 * Uri this driver should connect to. The driver supports bolt or neo4j as
+	 * schemes.
 	 */
 	private URI uri = URI.create("bolt://localhost:7687");
 
 	/**
-	 * The authentication the driver is supposed to use. Maybe null.
+	 * Authentication the driver is supposed to use. Maybe null.
 	 */
 	private Authentication authentication = new Authentication();
 
 	/**
-	 * The configuration of the connection pool.
+	 * Configuration of the connection pool.
 	 */
 	private PoolSettings pool = new PoolSettings();
 
@@ -95,52 +85,25 @@ public class Neo4jDriverProperties {
 		this.config = config;
 	}
 
-	public AuthToken getAuthToken() {
-		return this.authentication.asAuthToken();
-	}
-
-	public Config asDriverConfig() {
-
-		Config.ConfigBuilder builder = Config.builder();
-		this.pool.applyTo(builder);
-		String scheme = (this.uri != null) ? this.uri.getScheme() : "bolt";
-		this.config.applyTo(builder, isSimpleScheme(scheme));
-
-		return builder.withLogging(new Neo4jSpringJclLogging()).build();
-	}
-
-	static boolean isSimpleScheme(String scheme) {
-
-		String lowerCaseScheme = scheme.toLowerCase(Locale.ENGLISH);
-		try {
-			Scheme.validateScheme(lowerCaseScheme);
-		}
-		catch (IllegalArgumentException ex) {
-			throw new IllegalArgumentException(String.format("'%s' is not a supported scheme.", scheme));
-		}
-
-		return lowerCaseScheme.equals("bolt") || lowerCaseScheme.equals("neo4j");
-	}
-
 	public static class Authentication {
 
 		/**
-		 * The login of the user connecting to the database.
+		 * Login of the user connecting to the database.
 		 */
 		private String username;
 
 		/**
-		 * The password of the user connecting to the database.
+		 * Password of the user connecting to the database.
 		 */
 		private String password;
 
 		/**
-		 * The realm to connect to.
+		 * Realm to connect to.
 		 */
 		private String realm;
 
 		/**
-		 * A kerberos ticket for connecting to the database. Mutual exclusive with a given
+		 * Kerberos ticket for connecting to the database. Mutual exclusive with a given
 		 * username.
 		 */
 		private String kerberosTicket;
@@ -177,29 +140,6 @@ public class Neo4jDriverProperties {
 			this.kerberosTicket = kerberosTicket;
 		}
 
-		AuthToken asAuthToken() {
-
-			boolean hasUsername = StringUtils.hasText(this.username);
-			boolean hasPassword = StringUtils.hasText(this.password);
-			boolean hasKerberosTicket = StringUtils.hasText(this.kerberosTicket);
-
-			if (hasUsername && hasKerberosTicket) {
-				throw new InvalidConfigurationPropertyValueException("org.neo4j.driver.authentication",
-						"username=" + this.username + ",kerberos-ticket=" + this.kerberosTicket,
-						"Cannot specify both username and kerberos ticket.");
-			}
-
-			if (hasUsername && hasPassword) {
-				return AuthTokens.basic(this.username, this.password, this.realm);
-			}
-
-			if (hasKerberosTicket) {
-				return AuthTokens.kerberos(this.kerberosTicket);
-			}
-
-			return AuthTokens.none();
-		}
-
 	}
 
 	public static class PoolSettings {
@@ -215,7 +155,7 @@ public class Neo4jDriverProperties {
 		private boolean logLeakedSessions = false;
 
 		/**
-		 * The maximum amount of connections in the connection pool towards a single
+		 * Maximum amount of connections in the connection pool towards a single
 		 * database.
 		 */
 		private int maxConnectionPoolSize = org.neo4j.driver.internal.async.pool.PoolSettings.DEFAULT_MAX_CONNECTION_POOL_SIZE;
@@ -288,28 +228,6 @@ public class Neo4jDriverProperties {
 			this.metricsEnabled = metricsEnabled;
 		}
 
-		private void applyTo(Config.ConfigBuilder builder) {
-
-			if (this.logLeakedSessions) {
-				builder.withLeakedSessionsLogging();
-			}
-			builder.withMaxConnectionPoolSize(this.maxConnectionPoolSize);
-			if (this.idleTimeBeforeConnectionTest != null) {
-				builder.withConnectionLivenessCheckTimeout(this.idleTimeBeforeConnectionTest.toMillis(),
-						TimeUnit.MILLISECONDS);
-			}
-			builder.withMaxConnectionLifetime(this.maxConnectionLifetime.toMillis(), TimeUnit.MILLISECONDS);
-			builder.withConnectionAcquisitionTimeout(this.connectionAcquisitionTimeout.toMillis(),
-					TimeUnit.MILLISECONDS);
-
-			if (this.metricsEnabled) {
-				builder.withDriverMetrics();
-			}
-			else {
-				builder.withoutDriverMetrics();
-			}
-		}
-
 	}
 
 	public static class DriverSettings {
@@ -321,7 +239,7 @@ public class Neo4jDriverProperties {
 
 		/**
 		 * Specify how to determine the authenticity of an encryption certificate provided
-		 * by the Neo4j instance we are connecting to. Defaults to trust all.
+		 * by the Neo4j instance we are connecting to.
 		 */
 		private TrustSettings trustSettings = new TrustSettings();
 
@@ -382,30 +300,6 @@ public class Neo4jDriverProperties {
 			this.serverAddressResolverClass = serverAddressResolverClass;
 		}
 
-		private void applyTo(Config.ConfigBuilder builder, boolean withEncryptionAndTrustSettings) {
-
-			if (withEncryptionAndTrustSettings) {
-				applyEncryptionAndTrustSettings(builder);
-			}
-
-			builder.withConnectionTimeout(this.connectionTimeout.toMillis(), TimeUnit.MILLISECONDS);
-			builder.withMaxTransactionRetryTime(this.maxTransactionRetryTime.toMillis(), TimeUnit.MILLISECONDS);
-
-			if (this.serverAddressResolverClass != null) {
-				builder.withResolver(BeanUtils.instantiateClass(this.serverAddressResolverClass));
-			}
-		}
-
-		private void applyEncryptionAndTrustSettings(Config.ConfigBuilder builder) {
-			if (this.encrypted) {
-				builder.withEncryption();
-			}
-			else {
-				builder.withoutEncryption();
-			}
-			builder.withTrustStrategy(this.trustSettings.toInternalRepresentation());
-		}
-
 	}
 
 	public static class TrustSettings {
@@ -426,7 +320,7 @@ public class Neo4jDriverProperties {
 		private Strategy strategy = Strategy.TRUST_SYSTEM_CA_SIGNED_CERTIFICATES;
 
 		/**
-		 * The file of the certificate to use.
+		 * File of the certificate to use.
 		 */
 		private File certFile;
 
@@ -457,39 +351,6 @@ public class Neo4jDriverProperties {
 
 		public void setHostnameVerificationEnabled(boolean hostnameVerificationEnabled) {
 			this.hostnameVerificationEnabled = hostnameVerificationEnabled;
-		}
-
-		Config.TrustStrategy toInternalRepresentation() {
-			String propertyName = "org.neo4j.driver.config.trust-settings";
-
-			Config.TrustStrategy internalRepresentation;
-			switch (this.strategy) {
-			case TRUST_ALL_CERTIFICATES:
-				internalRepresentation = Config.TrustStrategy.trustAllCertificates();
-				break;
-			case TRUST_SYSTEM_CA_SIGNED_CERTIFICATES:
-				internalRepresentation = Config.TrustStrategy.trustSystemCertificates();
-				break;
-			case TRUST_CUSTOM_CA_SIGNED_CERTIFICATES:
-				if (this.certFile == null || !this.certFile.isFile()) {
-					throw new InvalidConfigurationPropertyValueException(propertyName, this.strategy.name(),
-							"Configured trust strategy requires a certificate file.");
-				}
-				internalRepresentation = Config.TrustStrategy.trustCustomCertificateSignedBy(this.certFile);
-				break;
-			default:
-				throw new InvalidConfigurationPropertyValueException(propertyName, this.strategy.name(),
-						"Unknown strategy.");
-			}
-
-			if (this.hostnameVerificationEnabled) {
-				internalRepresentation.withHostnameVerification();
-			}
-			else {
-				internalRepresentation.withoutHostnameVerification();
-			}
-
-			return internalRepresentation;
 		}
 
 	}
