+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/neo4j/Neo4jAutoConfigurationTests.java
@@ -35,10 +35,10 @@ import static org.mockito.Mockito.when;
 /**
  * @author Michael J. Simons
  */
-class Neo4jDriverAutoConfigurationTests {
+class Neo4jAutoConfigurationTests {
 
 	private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
-			.withConfiguration(AutoConfigurations.of(Neo4jDriverAutoConfiguration.class));
+			.withConfiguration(AutoConfigurations.of(Neo4jAutoConfiguration.class));
 
 	@Test
 	void shouldRequireAllNeededClasses() {
